+++ src/level_tester/api/app.py
@@ -121,7 +121,7 @@ async def lifespan(application: FastAPI):
     yield
 
 
-app = FastAPI(title="Levels Tester", version="0.4.4", lifespan=lifespan)
+app = FastAPI(title="Levels Tester", version="0.4.5", lifespan=lifespan)
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["http://127.0.0.1:8080", "http://localhost:8080"],